Move j.l.r.code.impl.OpWriter to j.l.r.code.writer
