+++ test/jdk/TEST.groups
@@ -651,3 +651,8 @@ jdk_containers_extended = \
 jdk_core_no_security = \
    :jdk_core \
    -:jdk_security
+
+
+jdk_lang_reflect_code = \
+   java/lang/reflect/code
+
